+++ CHANGELOG.md
@@ -643,6 +643,10 @@ $$Pow({x}, {n})=Pow\left({x}^2, \dfrac{n}{2}\right) * Pow\left(x, n \mod 2\right
 
 - Linary
 
+## 0172 - Factorial Trailing Zeroes
+
+- Count factor 2 & 5
+
 ## 0415 - Add strings
 
 - Full adder